ProjetDjango/requirements.txt
Django==5.1.6
django_phonenumber_field==8.0.0
django_rest_swagger==2.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
